rework networking with swift concurrency
